byte->sbyte (sbyte) casts at 5 sbyte-assignment sites: PlayerAccountData(GetEquipmentSlotType)/PlayerRestrictions(GetMaxLevelRestrict)/CMT_CHARACTER_INFORMATION x3(ReadC) (Java byte=signed=sbyte) (973->968)

+++ dotnetConversion/src/Aion.GameServer/Model/Account/PlayerAccountData.cs
@@ -108,7 +108,7 @@ public class PlayerAccountData
         List<VisibleItem> items = new();
         foreach (Item item in equipment)
         {
-            sbyte slotType = ItemSlotExtensions.GetEquipmentSlotType(item.GetEquipmentSlot());
+            sbyte slotType = (sbyte)ItemSlotExtensions.GetEquipmentSlotType(item.GetEquipmentSlot());
             if (slotType != 0)
                 items.Add(new VisibleItem(slotType, item.GetItemSkinTemplate().GetTemplateId(), item.GetGodStoneId(), item.GetItemColor()));
         }
